drm/lima: fix "still active bo inside vm"
This happens for exported bo when on screen rendering.
When the lima drm device fd is closing, the
lima_gem_free_object is not really called for the
exported bo because the display drm hold a reference
for it, but the lima vm is freed so need to remove
the va for this vm registered for this bo.

Signed-off-by: Qiang Yu <yuq825@gmail.com>